move entities to game folder behind feature flag
